Missing javadoc comments, missing user guide, user has to create a data directory for save system to work
Add javadoc comments to all non self-explanatory methods and classes
Add userguide
Storage automatically creates a data directory
