+++ src/main/java/lulu/Main.java
@@ -14,6 +14,14 @@ import javafx.stage.Stage;
 public class Main extends Application {
     private Lulu lulu = new Lulu();
 
+    /**
+     * Starts the application.
+     *
+     * @param stage the primary stage for this application, onto which
+     *              the application scene can be set.
+     *              Applications may create other stages, if needed, but they will not be
+     *              primary stages.
+     */
     @Override
     public void start(Stage stage) {
         try {
