+++ src/main/java/lulu/exception/LuluException.java
@@ -1,4 +1,8 @@
 package lulu.exception;
+
+/**
+ * The general exception to be thrown for the chatbot, Lulu.
+ */
 public class LuluException extends Exception {
     public LuluException(String message) {
         super(message);
